requirements.txt
streamlit
pandas
numpy
scipy
seaborn
yfinance
plotly
requests
python-dateutil
